tests/requirements.txt
selenium>=4.15.0
behave>=1.2.6
allure-behave>=2.13.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
